Refatorando o codigo para transformar a implementacao de gerar token em um modulo.
