+++ .idea/compiler.xml
@@ -7,10 +7,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="core" />
-        <module name="course" />
         <module name="gateway" />
+        <module name="token" />
         <module name="auth" />
+        <module name="course" />
+        <module name="core" />
         <module name="discovery" />
       </profile>
     </annotationProcessing>
@@ -21,6 +22,7 @@
       <module name="discovery" target="1.8" />
       <module name="gateway" target="1.8" />
       <module name="spring-boot-microservices-devdojo" target="1.8" />
+      <module name="token" target="1.8" />
     </bytecodeTargetLevel>
   </component>
   <component name="JavacSettings">
@@ -30,6 +32,7 @@
       <module name="course" options="-parameters" />
       <module name="discovery" options="-parameters" />
       <module name="gateway" options="-parameters" />
+      <module name="token" options="-parameters" />
     </option>
   </component>
 </project>
