+++ .idea/encodings.xml
@@ -7,5 +7,6 @@
     <file url="file://$PROJECT_DIR$/course/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/discovery/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/gateway/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/token/src/main/java" charset="UTF-8" />
   </component>
 </project>
