+++ auth/src/main/java/academy/devdojo/youtube/auth/security/config/SecurityCredentialsConfig.java
@@ -2,45 +2,34 @@ package academy.devdojo.youtube.auth.security.config;
 
 import academy.devdojo.youtube.auth.security.filter.JwtUsernameAndPasswordAuthenticationFilter;
 import academy.devdojo.youtube.core.property.JwtConfiguration;
+import academy.devdojo.youtube.security.config.SecurityTokenConfig;
+import academy.devdojo.youtube.security.token.creator.TokenCreator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.cors.CorsConfiguration;
-
-import javax.servlet.http.HttpServletResponse;
 
 @EnableWebSecurity
-public class SecurityCredentialsConfig extends WebSecurityConfigurerAdapter {
-    @Autowired
-    private UserDetailsService userDetailsService;
-    @Autowired
-    private JwtConfiguration jwtConfiguration;
+public class SecurityCredentialsConfig extends SecurityTokenConfig {
+    private final UserDetailsService userDetailsService;
+    private final TokenCreator tokenCreator;
+
+    public SecurityCredentialsConfig(JwtConfiguration jwtConfiguration,
+                                     UserDetailsService userDetailsService,
+                                     TokenCreator tokenCreator) {
+
+        this.userDetailsService = userDetailsService;
+        this.tokenCreator = tokenCreator;
+    }
 
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http
-                .csrf()
-                .disable()
-                .cors()
-                .configurationSource(request -> new CorsConfiguration().applyPermitDefaultValues())
-                .and()
-                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                .and()
-                .exceptionHandling().authenticationEntryPoint((req, resp, e) -> resp.sendError(
-                            HttpServletResponse.SC_UNAUTHORIZED))
-                .and()
-                .addFilter(new JwtUsernameAndPasswordAuthenticationFilter(authenticationManager(), jwtConfiguration))
-                .authorizeRequests()
-                .antMatchers(jwtConfiguration.getLoginUrl()).permitAll()
-                .antMatchers("/course/admin/**")
-                .hasRole("ADMIN")
-                .anyRequest().authenticated();
-
+                .addFilter(new JwtUsernameAndPasswordAuthenticationFilter(authenticationManager(), jwtConfiguration, tokenCreator));
+        super.configure(http);
     }
 
     @Override
