+++ auth/src/main/java/academy/devdojo/youtube/auth/security/filter/JwtUsernameAndPasswordAuthenticationFilter.java
@@ -2,15 +2,11 @@ package academy.devdojo.youtube.auth.security.filter;
 
 import academy.devdojo.youtube.core.model.ApplicationUser;
 import academy.devdojo.youtube.core.property.JwtConfiguration;
+import academy.devdojo.youtube.security.token.creator.TokenCreator;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nimbusds.jose.*;
-import com.nimbusds.jose.crypto.DirectEncrypter;
-import com.nimbusds.jose.crypto.RSASSASigner;
-import com.nimbusds.jose.jwk.JWK;
-import com.nimbusds.jose.jwk.RSAKey;
-import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
 import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +14,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
@@ -27,22 +22,18 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.security.KeyPair;
-import java.security.KeyPairGenerator;
-import java.security.interfaces.RSAPublicKey;
 import java.util.Collections;
-import java.util.Date;
-import java.util.UUID;
-import java.util.stream.Collectors;
 
-@AllArgsConstructor
+@RequiredArgsConstructor
 @Slf4j
 public class JwtUsernameAndPasswordAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 
     @Autowired
-    private AuthenticationManager authenticationManager;
+    private final AuthenticationManager authenticationManager;
     @Autowired
-    private JwtConfiguration jwtConfiguration;
+    private final JwtConfiguration jwtConfiguration;
+    @Autowired
+    private final TokenCreator tokenCreator;
 
     @SneakyThrows
     @Override
@@ -65,61 +56,9 @@ public class JwtUsernameAndPasswordAuthenticationFilter extends UsernamePassword
     @Override
     protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication auth) throws IOException, ServletException {
         log.info("Authentication was successfil for the user '{}', generating JWT token.");
-        SignedJWT signedJWT = createSignedJWT(auth);
-        String encryptedToken = encryptToken(signedJWT);
+        SignedJWT signedJWT = tokenCreator.createSignedJWT(auth);
+        String encryptedToken = tokenCreator.encryptToken(signedJWT);
         response.addHeader("Access-Control-Expose-Headers", "XSRF-TOKEN, "+ jwtConfiguration.getHeader().getName());
         response.addHeader(jwtConfiguration.getHeader().getName(), jwtConfiguration.getHeader().getPrefix() + encryptedToken);
-    }
-
-    @SneakyThrows
-    private SignedJWT createSignedJWT(Authentication auth){
-        ApplicationUser applicationUser = (ApplicationUser) auth.getPrincipal();
-        JWTClaimsSet jwtClaimsSet = createJWTClaimsSet(auth, applicationUser);
-        KeyPair rsaKeys = generateKeyPair();
-
-        JWK jwk = new RSAKey.Builder((RSAPublicKey) rsaKeys.getPublic()).keyID(UUID.randomUUID().toString()).build();
-
-        SignedJWT signedJWT = new SignedJWT(new JWSHeader.Builder(JWSAlgorithm.RS256)
-        .jwk(jwk)
-        .type(JOSEObjectType.JWT)
-        .build(), jwtClaimsSet);
-
-        RSASSASigner signer = new RSASSASigner(rsaKeys.getPrivate());
-        signedJWT.sign(signer);
-
-        log.info("Token assinado '{}'", signedJWT.serialize());
-        return signedJWT;
-    }
-
-    private JWTClaimsSet createJWTClaimsSet(Authentication auth, ApplicationUser applicationUser){
-        return new JWTClaimsSet.Builder()
-                .subject(applicationUser.getUsername())
-                .claim("authorities", auth.getAuthorities()
-                .stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.toList()))
-                .issuer("http://academy.devdojo")
-                .issueTime(new Date())
-                .expirationTime(new Date(System.currentTimeMillis() + (jwtConfiguration.getExpiration() * 1000)))
-                .build();
-    }
-
-    @SneakyThrows
-    private KeyPair generateKeyPair(){
-        KeyPairGenerator generator = KeyPairGenerator.getInstance("RSA");
-        generator.initialize(2048);
-        return generator.genKeyPair();
-    }
-
-    private String encryptToken(SignedJWT signedJWT) throws JOSEException {
-        DirectEncrypter directEncrypter = new DirectEncrypter(jwtConfiguration.getPrivateKey().getBytes());
-
-        JWEObject jweObject = new JWEObject(new JWEHeader.Builder(JWEAlgorithm.DIR, EncryptionMethod.A128CBC_HS256)
-        .contentType("JWT")
-        .build(), new Payload(signedJWT));
-
-        jweObject.encrypt(directEncrypter);
-
-        return jweObject.serialize();
     }
 }
