+++ auth/src/main/resources/application.yml
@@ -7,12 +7,11 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:8081/eureka/
-    fetch-registry: true
     register-with-eureka: true
 
 spring:
   application:
-    name: /auth
+    name: auth
   jpa:
     show-sql: false
     hibernate:
