+++ token/src/main/java/academy/devdojo/youtube/security/config/SecurityTokenConfig.java
@@ -1,0 +1,35 @@
+package academy.devdojo.youtube.security.config;
+
+import academy.devdojo.youtube.core.property.JwtConfiguration;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.web.cors.CorsConfiguration;
+
+import javax.servlet.http.HttpServletResponse;
+
+public class SecurityTokenConfig extends WebSecurityConfigurerAdapter {
+    @Autowired
+    protected JwtConfiguration jwtConfiguration;
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http
+                .csrf()
+                .disable()
+                .cors()
+                .configurationSource(request -> new CorsConfiguration().applyPermitDefaultValues())
+                .and()
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .and()
+                .exceptionHandling().authenticationEntryPoint((req, resp, e) -> resp.sendError(
+                            HttpServletResponse.SC_UNAUTHORIZED))
+                .and()
+                .authorizeRequests()
+                .antMatchers(jwtConfiguration.getLoginUrl()).permitAll()
+                .antMatchers("/course/admin/**")
+                .hasRole("ADMIN")
+                .anyRequest().authenticated();
+    }
+}
