+++ token/src/main/java/academy/devdojo/youtube/security/token/creator/TokenCreator.java
@@ -1,0 +1,86 @@
+package academy.devdojo.youtube.security.token.creator;
+
+import academy.devdojo.youtube.core.model.ApplicationUser;
+import academy.devdojo.youtube.core.property.JwtConfiguration;
+import com.nimbusds.jose.*;
+import com.nimbusds.jose.crypto.DirectEncrypter;
+import com.nimbusds.jose.crypto.RSASSASigner;
+import com.nimbusds.jose.jwk.JWK;
+import com.nimbusds.jose.jwk.RSAKey;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
+import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.stereotype.Service;
+
+import java.security.KeyPair;
+import java.security.KeyPairGenerator;
+import java.security.interfaces.RSAPublicKey;
+import java.util.Date;
+import java.util.UUID;
+import java.util.stream.Collectors;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class TokenCreator {
+    @Autowired
+    private JwtConfiguration jwtConfiguration;
+
+    @SneakyThrows
+    public SignedJWT createSignedJWT(Authentication auth){
+        ApplicationUser applicationUser = (ApplicationUser) auth.getPrincipal();
+        JWTClaimsSet jwtClaimsSet = createJWTClaimsSet(auth, applicationUser);
+        KeyPair rsaKeys = generateKeyPair();
+
+        JWK jwk = new RSAKey.Builder((RSAPublicKey) rsaKeys.getPublic()).keyID(UUID.randomUUID().toString()).build();
+
+        SignedJWT signedJWT = new SignedJWT(new JWSHeader.Builder(JWSAlgorithm.RS256)
+                .jwk(jwk)
+                .type(JOSEObjectType.JWT)
+                .build(), jwtClaimsSet);
+
+        RSASSASigner signer = new RSASSASigner(rsaKeys.getPrivate());
+        signedJWT.sign(signer);
+
+        log.info("Token assinado '{}'", signedJWT.serialize());
+        return signedJWT;
+    }
+
+    private JWTClaimsSet createJWTClaimsSet(Authentication auth, ApplicationUser applicationUser){
+        return new JWTClaimsSet.Builder()
+                .subject(applicationUser.getUsername())
+                .claim("authorities", auth.getAuthorities()
+                        .stream()
+                        .map(GrantedAuthority::getAuthority)
+                        .collect(Collectors.toList()))
+                .issuer("http://academy.devdojo")
+                .issueTime(new Date())
+                .expirationTime(new Date(System.currentTimeMillis() + (jwtConfiguration.getExpiration() * 1000)))
+                .build();
+    }
+
+    @SneakyThrows
+    private KeyPair generateKeyPair(){
+        KeyPairGenerator generator = KeyPairGenerator.getInstance("RSA");
+        generator.initialize(2048);
+        return generator.genKeyPair();
+    }
+
+    public String encryptToken(SignedJWT signedJWT) throws JOSEException {
+        DirectEncrypter directEncrypter = new DirectEncrypter(jwtConfiguration.getPrivateKey().getBytes());
+
+        JWEObject jweObject = new JWEObject(new JWEHeader.Builder(JWEAlgorithm.DIR, EncryptionMethod.A128CBC_HS256)
+                .contentType("JWT")
+                .build(), new Payload(signedJWT));
+
+        jweObject.encrypt(directEncrypter);
+
+        return jweObject.serialize();
+    }
+}
